tell useItemSubscrption's effect to rerun on id change
* fixes bug where the card components were not updating, even when the
data passed to them as props had changed. Useeffect wasnt running again,
because I had passed it an empty array telling it to not rerun. I needed
it to set up the effect again when the ID it was passed changed, because
the state is updated in that effect. Without the effect running, the
state wasn't switching to the new ID.